Confusion with speech marks on line 19 in script.js
A back slash used as an escape character for the apostrophe in (It's) so
the script doesn't think it is the end of the quotation. All other
speech marks in this line have been change to single speech marks to
avoid confusion with double speech marks in the middle of a quotation to
avoid the same problem.