randomBackgroundColour not defined on line 12 in index.html
Have removed the whole startTimer function as requested in the comment,
this will resolve the problem of randomBackgroundColour not being
defined as it no longer exists. This also resolves a possible future
problem as there is a startTimer function in script.js which wouldn't
have been called onload (as in line 17 of index.html), the startTimer
function in index.html would've been called instead. Now the script.js
startTimer function will load instead.