purchase rate changed on line 49 of script.java
the purchase rate at 1000 meant that the offer given was 100 times too
big, they would be paying more than the value of the car was originally
which makes no sense. Have changed this to 10 as requested, this way it
produces an offer where they can make profit. Spelling mistake changed
on lines 49 and 51 from pruchaseRate to purchaseRate.